CMS-577: Refactor with flash and confirms
