+++ frontend/src/router/pages/PreviewWinterFeesChanges.jsx
@@ -1,10 +1,7 @@
 import PropTypes from "prop-types";
 import { Link, useOutletContext } from "react-router-dom";
 import { useApiPost } from "@/hooks/useApi";
-import { useNavigationGuard } from "@/hooks/useNavigationGuard";
-import { useConfirmation } from "@/hooks/useConfirmation";
 import { useMissingDatesConfirmation } from "@/hooks/useMissingDatesConfirmation";
-import { useFlashMessage } from "@/hooks/useFlashMessage";
 import paths from "@/router/paths";
 
 import { faPen } from "@fa-kit/icons/classic/solid";
@@ -16,9 +13,7 @@ import ContactBox from "@/components/ContactBox";
 import ReadyToPublishBox from "@/components/ReadyToPublishBox";
 import DateRange from "@/components/DateRange";
 import ChangeLogsList from "@/components/ChangeLogsList";
-import ConfirmationDialog from "@/components/ConfirmationDialog";
 import MissingDatesConfirmationDialog from "@/components/MissingDatesConfirmationDialog";
-import FlashMessage from "@/components/FlashMessage";
 
 import "./PreviewChanges.scss";
 
@@ -34,31 +29,16 @@ function PreviewChanges() {
     navigate,
     navigateAndScroll,
     saveAsDraft,
-    hasChanges,
     saving,
+    hasChanges,
+    showErrorFlash,
   } = useOutletContext();
 
   const { sendData: approveData, loading: savingApproval } = useApiPost(
     `/seasons/${seasonId}/approve/`,
   );
 
-  const errorFlashMessage = useFlashMessage();
-
-  const {
-    title,
-    message,
-    confirmationDialogNotes,
-    confirmButtonText,
-    cancelButtonText,
-    openConfirmation,
-    handleConfirm,
-    handleCancel,
-    isConfirmationOpen,
-  } = useConfirmation();
-
   const missingDatesConfirmation = useMissingDatesConfirmation();
-
-  useNavigationGuard(hasChanges, openConfirmation);
 
   function navigateToEdit() {
     navigateAndScroll(paths.winterFeesEdit(parkId, seasonId));
@@ -76,13 +56,6 @@ function PreviewChanges() {
         end={date.endDate}
       />
     ));
-  }
-
-  function showErrorFlash() {
-    errorFlashMessage.openFlashMessage(
-      "Unable to save changes",
-      "There was a problem saving these changes. Please try again.",
-    );
   }
 
   function getFeaturesWithMissingDates() {
@@ -194,25 +167,6 @@ function PreviewChanges() {
   return (
     <div className="container">
       <div className="page review-winter-fees-changes">
-        <FlashMessage
-          title={errorFlashMessage.flashTitle}
-          message={errorFlashMessage.flashMessage}
-          isVisible={errorFlashMessage.isFlashOpen}
-          onClose={errorFlashMessage.handleFlashClose}
-          variant="error"
-        />
-
-        <ConfirmationDialog
-          title={title}
-          message={message}
-          notes={confirmationDialogNotes}
-          confirmButtonText={confirmButtonText}
-          cancelButtonText={cancelButtonText}
-          onCancel={handleCancel}
-          onConfirm={handleConfirm}
-          isOpen={isConfirmationOpen}
-        />
-
         <MissingDatesConfirmationDialog
           featureNames={missingDatesConfirmation.featureNames}
           inputMessage={missingDatesConfirmation.inputMessage}
@@ -291,7 +245,7 @@ function PreviewChanges() {
           <button
             type="button"
             className="btn btn-outline-primary"
-            onClick={() => saveAsDraft(openConfirmation, showErrorFlash)}
+            onClick={saveAsDraft}
             disabled={!hasChanges()}
           >
             Save draft
